ml_service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
scikit-learn==1.5.0
xgboost==2.0.3
pandas==2.2.2
numpy==1.26.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlalchemy==2.0.30
joblib==1.4.2
pydantic==2.7.1